+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the `luamml` package since the
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 this project uses date-based 'snapshot' version identifiers.
+## [Unreleased]
+### Changed
+  * luamml-amsmath.lua: extended add_intent_pause to add the :pause-medium intent also
+  to mathml-SE
+  * luamml.dtx: predeclare the intent-pause-medium attribute.
+  * luamml-structelemwriter.lua: change key newattribute to role/new-attribute.
 
 ## [2025-06-25]
  ### Fixed